chore: remove pieces file
